add build.

+++ scripts/entrypoint.sh
@@ -3,24 +3,8 @@
 set -o errexit
 set -o pipefail
 
-if [ -z "$WEEWX_CONFIG_FILE" ]; then
+if [ -z "${CONFIG_FILE}" ]; then
 	export CONF_FILE="/data/weewx.conf"
-fi
-
-# due to the way weewx generates and processes data from remotes
-# we need to ensure the user has set their timezone correctly (esp if migrated)
-if [ -z "$TZ" ]; then 
-	echo "timezone not set"
-	exit 1
-fi
-
-# set timezone
-#cp "/usr/share/zoneinfo/${TZ}" /etc/localtime && echo "${TZ}" > /etc/timezone
-
-if [ ! -f "${CONF_FILE}" ]; then
-	echo "error! no configuration file"
-	logger "error, no configuration file"
-	exit 1
 fi
 
 if [ "$1" = "--version" ]; then
@@ -36,6 +20,21 @@ elif [ "$1" = "--init" ]; then
 	wee_config --intall "${CONF_FILE}"
 	exit $?
 elif [ "$1" = "--run" ]; then
+	# due to the way weewx generates and processes data from remotes
+	# we need to ensure the user has set their timezone correctly (esp if migrated)
+	if [ -z "$TZ" ]; then 
+		echo "timezone not set"
+		exit 1
+	fi
+	
+	# set timezone
+	#cp "/usr/share/zoneinfo/${TZ}" /etc/localtime && echo "${TZ}" > /etc/timezone
+	
+	if [ ! -f "${CONF_FILE}" ]; then
+		echo "error! no configuration file"
+		logger "error, no configuration file"
+		exit 1
+	fi
 	# as a backup for plugins that hard code the config path
 	# we put it in the etc directoruy, but use the /data path
 	cp -v /data/weewx.conf /etc/weewx/weewx.conf
